Formats/Image: - Fixed images not working
Formats/TIKI:
- Fixed predefined negative frames not taken into account

Formats/Skel:
- Fixed major issues with unmatched bone channels
- Fixed channels being invalid on old skeleton animations

Managers/SoundManager:
- Added support for instructions 'settiki' and 'end'
- Fixed bad parse behavior due to non-implemented parameters 'always', 'forcesubtitle' and 'maps'
- Allow multiple lines inside strings

Script:
- Added a parameter to allow/prevent multiple lines in Script::GetString

+++ tests/Level.cpp
@@ -195,7 +195,7 @@ public:
 		return "BSP";
 	}
 
-	virtual unsigned int priority() override
+	virtual long priority() override
 	{
 		return 0;
 	}
